:recycle: refactor downcast trait
