+++ plugins/cp2077-rs/src/lib.rs
@@ -1,66 +1,17 @@
-mod game;
-use std::marker::PhantomData;
-
-pub use game::*;
-mod transaction;
-use red4ext_rs::{
-    prelude::RefRepr,
-    types::{IScriptable, Ref},
-};
-pub use transaction::*;
-mod time;
-pub use time::*;
-mod housing;
-pub use housing::*;
 mod board;
-pub use board::*;
-mod event;
-pub use event::*;
 mod delay;
+mod downcast;
+mod event;
+mod game;
+mod housing;
+mod time;
+mod transaction;
+
+pub use board::*;
 pub use delay::*;
-
-#[derive(Default, Clone)]
-#[repr(transparent)]
-pub struct TypedRef<T>(Ref<IScriptable>, PhantomData<T>);
-
-impl<T> TypedRef<T> {
-    pub fn new(reference: Ref<IScriptable>) -> Self {
-        Self(reference, PhantomData)
-    }
-}
-
-/// SAFETY: implementations of this trait are only valid if your implementors are indeed child classes of `Parent`
-pub unsafe trait IntoTypedRef<Parent: RefRepr + Default + Clone> {
-    /// cast a reference into a parent typed reference
-    fn into_typed_ref(self) -> TypedRef<Parent>;
-}
-
-pub trait FromTypedRef<Parent: RefRepr + Default + Clone>: Sized {
-    /// cast any self typed reference into itself
-    fn from_typed_ref(reference: TypedRef<Parent>) -> Self;
-}
-
-pub trait Downcast<Parent>: IntoTypedRef<Parent>
-where
-    Self: IntoTypedRef<Parent>,
-    Parent: RefRepr + Default + Clone + FromTypedRef<Parent>,
-{
-    /// automatically downcast a wrapper over Ref<IScriptable>
-    /// into its parent
-    fn downcast(self) -> Parent
-    where
-        Self: Sized,
-    {
-        let typed = self.into_typed_ref();
-        let parent = Parent::from_typed_ref(typed);
-        parent
-    }
-}
-
-/// automatically implements downcast for any child classes that can be turned into their parent class
-impl<Child, Parent> Downcast<Parent> for Child
-where
-    Self: IntoTypedRef<Parent>,
-    Parent: RefRepr + Default + Clone + FromTypedRef<Parent>,
-{
-}
+pub use downcast::*;
+pub use event::*;
+pub use game::*;
+pub use housing::*;
+pub use time::*;
+pub use transaction::*;
